a bit of rearranging, removing some unused error types, clarifying and adding documentation, and linkifying links.
