+++ Sources/OpenAPIKit/Content.swift
@@ -10,7 +10,7 @@ import Foundation
 extension OpenAPI {
     /// OpenAPI Spec "Media Type Object"
     /// 
-    /// https://github.com/OAI/OpenAPI-Specification/blob/master/versions/3.0.3.md#media-type-object
+    /// See [OpenAPI Media Type Object](https://github.com/OAI/OpenAPI-Specification/blob/master/versions/3.0.3.md#media-type-object).
     public struct Content: Equatable, CodableVendorExtendable {
         public var schema: Either<JSONReference<JSONSchema>, JSONSchema>
         public var example: AnyCodable?
