+++ Sources/OpenAPIKit/Utility/Encodable+EncodeIfNotNil.swift
