+++ Sources/OpenAPIKit/Utility/Result+Value.swift
